+++ src/main/java/com/kruemblegard/client/render/CephalariRenderer.java
@@ -1,5 +1,6 @@
 package com.kruemblegard.client.render;
 
+import com.kruemblegard.client.render.layer.CephalariBodyOverlayLayer;
 import com.kruemblegard.client.render.layer.CephalariProfessionLayer;
 import com.kruemblegard.client.render.model.CephalariModel;
 import com.kruemblegard.entity.CephalariEntity;
@@ -13,6 +14,7 @@ public class CephalariRenderer extends GeoEntityRenderer<CephalariEntity> {
         super(renderManager, new CephalariModel());
         this.shadowRadius = 0.5F;
 
+        addRenderLayer(new CephalariBodyOverlayLayer(this));
         addRenderLayer(new CephalariProfessionLayer(this));
     }
 }